Restructure as copier template with customizable Python scripts
Major refactoring from Python package to copier template approach:

TEMPLATE STRUCTURE:
- Repository root IS the copier template
- copier.yml at root with interactive questions
- All Python scripts, Jinja templates, and LaTeX files get copied to projects
- Users can freely edit Python code for customization

CHANGES:
- Moved all scripts to olypack/ directory in template
- Created olypack.py CLI for cleaner Makefile commands
- Added encrypt_pdf function to utils.py
- Makefile now uses clean commands like:
  - python3 olypack.py encrypt INPUT.pdf OUTPUT.pdf password
  - python3 olypack.py shuffle
  - python3 olypack.py packet

MINIMAL PACKAGE:
- Kept minimal olypack package (src/) for install command only
- olypack install: downloads LaTeX/Asy files, sets up prek hooks
- Install via: uv tool install git+https://github.com/vEnhance/olypack

WORKFLOW:
1. Install olypack tool once: uv tool install git+https://github.com/vEnhance/olypack
2. Create project: copier copy gh:vEnhance/olypack my-test-2026
3. Setup dependencies: olypack install
4. Customize Python scripts in olypack/ as needed
5. Use Makefile or olypack.py commands

This approach allows maintainers to edit Python scripts per-project
rather than having a black-box package they can't customize.
