+++ olypack/__init__.py
@@ -1,0 +1,1 @@
+# olypack utilities for managing olympiad problem packets
