+++ olypack/jinja-templates/author-table.tex.jinja
