+++ olypack/jinja-templates/data-solns.tex.jinja
