+++ olypack/jinja-templates/problems-day.tex.jinja
