+++ olypack/jinja-templates/solutions-day.tex.jinja
